# Add an SSE4.1 accelerated implementation of `Coder1248`