--- a/.cargo/config
+++ b/.cargo/config
@@ -1,2 +1,2 @@
 [target.'cfg(target_arch = "x86_64")']
-rustflags = ["-Ctarget-feature=+ssse3"]
+rustflags = ["-Ctarget-feature=+ssse3,+sse4.1"]