--- a/src/coder1248/mod.rs
+++ b/src/coder1248/mod.rs
@@ -1,7 +1,7 @@
 #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
 mod neon;
 #[cfg(target_arch = "x86_64")]
-mod ssse3;
+mod sse41;
 
 use crate::coding_descriptor::CodingDescriptor;
 use crate::{coder_impl, Coder};
@@ -50,7 +50,7 @@ enum Impl {
     Neon,
     #[cfg(all(target_arch = "x86_64"))]
     #[allow(dead_code)]
-    SSSE3,
+    SSE41,
 }
 
 /// `Coder1248` packs 64-bit integers into lengths of 1, 2, 4, or 8 bytes.
@@ -69,10 +69,14 @@ impl Coder for Coder1248 {
                 return Coder1248(Impl::Neon);
             }
         }
-        #[cfg(all(target_arch = "x86_64", target_feature = "ssse3"))]
+        #[cfg(all(
+            target_arch = "x86_64",
+            target_feature = "ssse3",
+            target_feature = "sse4.1"
+        ))]
         {
-            if std::arch::is_x86_feature_detected!("ssse3") {
-                return Coder1248(Impl::SSSE3);
+            if is_x86_feature_detected!("ssse3") && is_x86_feature_detected!("sse4.1") {
+                return Coder1248(Impl::SSE41);
             }
         }
         Coder1248(Impl::Scalar)
@@ -84,7 +88,7 @@ impl Coder for Coder1248 {
             #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
             Impl::Neon => coder_impl::encode::<neon::RawGroupImpl>(values, tags, encoded),
             #[cfg(target_arch = "x86_64")]
-            Impl::SSSE3 => coder_impl::encode::<ssse3::RawGroupImpl>(values, tags, encoded),
+            Impl::SSE41 => coder_impl::encode::<sse41::RawGroupImpl>(values, tags, encoded),
         }
     }
 
@@ -104,8 +108,8 @@ impl Coder for Coder1248 {
                 coder_impl::encode_deltas::<neon::RawGroupImpl>(initial, values, tags, encoded)
             }
             #[cfg(target_arch = "x86_64")]
-            Impl::SSSE3 => {
-                coder_impl::encode_deltas::<ssse3::RawGroupImpl>(initial, values, tags, encoded)
+            Impl::SSE41 => {
+                coder_impl::encode_deltas::<sse41::RawGroupImpl>(initial, values, tags, encoded)
             }
         }
     }
@@ -116,7 +120,7 @@ impl Coder for Coder1248 {
             #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
             Impl::Neon => coder_impl::decode::<neon::RawGroupImpl>(tags, encoded, values),
             #[cfg(target_arch = "x86_64")]
-            Impl::SSSE3 => coder_impl::decode::<ssse3::RawGroupImpl>(tags, encoded, values),
+            Impl::SSE41 => coder_impl::decode::<sse41::RawGroupImpl>(tags, encoded, values),
         }
     }
 
@@ -136,8 +140,8 @@ impl Coder for Coder1248 {
                 coder_impl::decode_deltas::<neon::RawGroupImpl>(initial, tags, encoded, values)
             }
             #[cfg(target_arch = "x86_64")]
-            Impl::SSSE3 => {
-                coder_impl::decode_deltas::<ssse3::RawGroupImpl>(initial, tags, encoded, values)
+            Impl::SSE41 => {
+                coder_impl::decode_deltas::<sse41::RawGroupImpl>(initial, tags, encoded, values)
             }
         }
     }
@@ -148,7 +152,7 @@ impl Coder for Coder1248 {
             #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
             Impl::Neon => coder_impl::data_len::<neon::RawGroupImpl>(tags),
             #[cfg(target_arch = "x86_64")]
-            Impl::SSSE3 => coder_impl::data_len::<ssse3::RawGroupImpl>(tags),
+            Impl::SSE41 => coder_impl::data_len::<sse41::RawGroupImpl>(tags),
         }
     }
 
@@ -158,7 +162,7 @@ impl Coder for Coder1248 {
             #[cfg(all(target_arch = "aarch64", target_endian = "little"))]
             Impl::Neon => coder_impl::skip_deltas::<neon::RawGroupImpl>(tags, encoded),
             #[cfg(target_arch = "x86_64")]
-            Impl::SSSE3 => coder_impl::skip_deltas::<ssse3::RawGroupImpl>(tags, encoded),
+            Impl::SSE41 => coder_impl::skip_deltas::<sse41::RawGroupImpl>(tags, encoded),
         }
     }
 }