This requires SSE4.1 instead of SSSE3 like `Coder1234` because of a few instructions used in `encode()`. Otherwise this
is pretty similar but requires dividing each group in half (2 instead of 4) in order to perform all operations ~doubling the
number of instructions run.

```
Coder1248/encode/1      time:   [688.43 ns 712.29 ns 742.73 ns]                                
                        thrpt:  [1.3787 Gelem/s 1.4376 Gelem/s 1.4874 Gelem/s]
                 change:
                        time:   [-59.800% -57.094% -54.556%] (p = 0.00 < 0.05)
                        thrpt:  [+120.05% +133.07% +148.76%]
                        Performance has improved.
Coder1248/encode_deltas/1                                                                             
                        time:   [846.81 ns 879.11 ns 918.99 ns]
                        thrpt:  [1.1143 Gelem/s 1.1648 Gelem/s 1.2092 Gelem/s]
                 change:
                        time:   [-60.406% -59.050% -57.418%] (p = 0.00 < 0.05)
                        thrpt:  [+134.84% +144.20% +152.57%]
                        Performance has improved.
Coder1248/decode/1      time:   [553.25 ns 569.95 ns 589.73 ns]                                
                        thrpt:  [1.7364 Gelem/s 1.7966 Gelem/s 1.8509 Gelem/s]
                 change:
                        time:   [-32.638% -29.495% -26.343%] (p = 0.00 < 0.05)
                        thrpt:  [+35.765% +41.835% +48.452%]
                        Performance has improved.
Coder1248/decode_deltas/1                                                                             
                        time:   [651.41 ns 673.40 ns 700.64 ns]
                        thrpt:  [1.4615 Gelem/s 1.5206 Gelem/s 1.5720 Gelem/s]
                 change:
                        time:   [-37.844% -33.409% -29.202%] (p = 0.00 < 0.05)
                        thrpt:  [+41.248% +50.171% +60.884%]
                        Performance has improved.
Coder1248/skip_deltas/1 time:   [542.16 ns 561.37 ns 587.57 ns]                                     
                        thrpt:  [1.7428 Gelem/s 1.8241 Gelem/s 1.8887 Gelem/s]
                 change:
                        time:   [-42.197% -39.683% -37.328%] (p = 0.00 < 0.05)
                        thrpt:  [+59.560% +65.790% +73.002%]
                        Performance has improved.
Coder1248/data_len/1    time:   [86.656 ns 88.934 ns 91.883 ns]                                 
                        thrpt:  [11.145 Gelem/s 11.514 Gelem/s 11.817 Gelem/s]
                 change:
                        time:   [-19.713% -15.441% -11.380%] (p = 0.00 < 0.05)
                        thrpt:  [+12.841% +18.261% +24.553%]
                        Performance has improved.
Coder1248/encode/4      time:   [708.16 ns 732.35 ns 763.91 ns]                                
                        thrpt:  [1.3405 Gelem/s 1.3982 Gelem/s 1.4460 Gelem/s]
                 change:
                        time:   [-76.996% -76.313% -75.636%] (p = 0.00 < 0.05)
                        thrpt:  [+310.45% +322.18% +334.71%]
                        Performance has improved.
Coder1248/encode_deltas/4                                                                             
                        time:   [885.67 ns 920.30 ns 960.93 ns]
                        thrpt:  [1.0656 Gelem/s 1.1127 Gelem/s 1.1562 Gelem/s]
                 change:
                        time:   [-65.619% -63.226% -59.395%] (p = 0.00 < 0.05)
                        thrpt:  [+146.28% +171.93% +190.86%]
                        Performance has improved.
Coder1248/decode/4      time:   [564.93 ns 592.38 ns 628.11 ns]                                
                        thrpt:  [1.6303 Gelem/s 1.7286 Gelem/s 1.8126 Gelem/s]
                 change:
                        time:   [-36.025% -30.883% -25.997%] (p = 0.00 < 0.05)
                        thrpt:  [+35.129% +44.681% +56.311%]
                        Performance has improved.
Coder1248/decode_deltas/4                                                                             
                        time:   [664.58 ns 689.44 ns 721.52 ns]
                        thrpt:  [1.4192 Gelem/s 1.4853 Gelem/s 1.5408 Gelem/s]
                 change:
                        time:   [-30.291% -27.265% -24.007%] (p = 0.00 < 0.05)
                        thrpt:  [+31.590% +37.485% +43.454%]
                        Performance has improved.
Coder1248/skip_deltas/4 time:   [565.87 ns 592.91 ns 637.67 ns]                                     
                        thrpt:  [1.6059 Gelem/s 1.7271 Gelem/s 1.8096 Gelem/s]
                 change:
                        time:   [-40.827% -36.088% -30.983%] (p = 0.00 < 0.05)
                        thrpt:  [+44.891% +56.464% +68.997%]
                        Performance has improved.
Coder1248/data_len/4    time:   [87.497 ns 91.482 ns 96.599 ns]                                 
                        thrpt:  [10.601 Gelem/s 11.194 Gelem/s 11.703 Gelem/s]
                 change:
                        time:   [-15.891% -12.125% -8.3843%] (p = 0.00 < 0.05)
                        thrpt:  [+9.1516% +13.798% +18.893%]
                        Performance has improved.
Coder1248/encode/8      time:   [691.03 ns 708.77 ns 730.76 ns]                                
                        thrpt:  [1.4013 Gelem/s 1.4448 Gelem/s 1.4818 Gelem/s]
                 change:
                        time:   [-74.611% -73.773% -72.896%] (p = 0.00 < 0.05)
                        thrpt:  [+268.95% +281.28% +293.88%]
                        Performance has improved.
Coder1248/encode_deltas/8                                                                             
                        time:   [898.42 ns 933.04 ns 973.62 ns]
                        thrpt:  [1.0517 Gelem/s 1.0975 Gelem/s 1.1398 Gelem/s]
                 change:
                        time:   [-66.055% -63.955% -61.969%] (p = 0.00 < 0.05)
                        thrpt:  [+162.94% +177.43% +194.60%]
                        Performance has improved.
Coder1248/decode/8      time:   [554.54 ns 571.99 ns 596.05 ns]                                
                        thrpt:  [1.7180 Gelem/s 1.7902 Gelem/s 1.8466 Gelem/s]
                 change:
                        time:   [-35.713% -32.679% -29.763%] (p = 0.00 < 0.05)
                        thrpt:  [+42.375% +48.542% +55.552%]
                        Performance has improved.
Coder1248/decode_deltas/8                                                                             
                        time:   [660.32 ns 688.83 ns 729.63 ns]
                        thrpt:  [1.4035 Gelem/s 1.4866 Gelem/s 1.5508 Gelem/s]
                 change:
                        time:   [-30.349% -27.541% -24.244%] (p = 0.00 < 0.05)
                        thrpt:  [+32.003% +38.010% +43.574%]
                        Performance has improved.
Coder1248/skip_deltas/8 time:   [560.27 ns 591.64 ns 636.58 ns]                                     
                        thrpt:  [1.6086 Gelem/s 1.7308 Gelem/s 1.8277 Gelem/s]
                 change:
                        time:   [-44.146% -40.720% -37.144%] (p = 0.00 < 0.05)
                        thrpt:  [+59.093% +68.691% +79.037%]
                        Performance has improved.
Coder1248/data_len/8    time:   [86.919 ns 91.829 ns 99.446 ns]                                 
                        thrpt:  [10.297 Gelem/s 11.151 Gelem/s 11.781 Gelem/s]
                 change:
                        time:   [-14.295% -9.5187% -3.7258%] (p = 0.00 < 0.05)
                        thrpt:  [+3.8699% +10.520% +16.679%]
                        Performance has improved.
```